Session storage now stores message conversation, message conversation displays sequentially on Display through ul list. Screen automatically adjusts to bottom of scroll (on reload, etc.).
