+++ client/src/lib/filterSetData.ts
@@ -1,0 +1,19 @@
+//will use
+export type UnsavedFilterSet = {
+  label: string;
+  person: boolean;
+  phoneNumber: boolean;
+  emailAddress: boolean;
+  dateTime: boolean;
+  location: boolean;
+  usSsn: boolean;
+  usDriverLicense: boolean;
+  crypto: boolean;
+  usBankNumber: boolean;
+  creditCard: boolean;
+  ipAddress: boolean;
+};
+//will use
+export type FilterSet = UnsavedFilterSet & {
+  filterSetId: number;
+};
