+++ client/src/lib/messageData.ts
@@ -1,0 +1,5 @@
+export type Message = {
+  id: number;
+  text: string;
+  sender: 'user' | 'redactedGpt' | 'loading';
+};
